requirements.txt
praw==7.7.1
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
lxml==4.9.3
nltk==3.8.1